ch10_full_pipeline/requirements.txt
# Minimal direct dependencies for the current codebase
# (transitive dependencies are resolved by pip)

typer==0.21.2
pydantic-settings==2.13.1
python-dotenv==1.2.1

datasets==4.4.0
sentence-transformers==5.2.3
faiss-cpu==1.12.0
numpy==2.2.6

ollama==0.6.1

SQLAlchemy==2.0.47
pgvector==0.4.2
psycopg2-binary==2.9.11

pytest==8.3.5
